review, reservation 빼고 완료

+++ src/main/java/com/jeong/studyroomreservation/web/controller/CompanyController.java
@@ -1,12 +1,13 @@
 package com.jeong.studyroomreservation.web.controller;
 
-import com.jeong.studyroomreservation.domain.dto.CompanyDto;
-import com.jeong.studyroomreservation.domain.dto.UserDto;
+import com.jeong.studyroomreservation.domain.dto.company.CompanyResponseDto;
+import com.jeong.studyroomreservation.domain.dto.company.CompanyUpdateResponseDto;
+import com.jeong.studyroomreservation.domain.dto.user.UserDto;
 import com.jeong.studyroomreservation.domain.entity.compnay.Company;
 import com.jeong.studyroomreservation.domain.entity.compnay.CompanyMapper;
 import com.jeong.studyroomreservation.domain.entity.user.UserRole;
 import com.jeong.studyroomreservation.domain.service.CompanyService;
-import com.jeong.studyroomreservation.web.dto.CompanyUpdateDto;
+import com.jeong.studyroomreservation.web.dto.company.CompanyUpdateDto;
 import com.jeong.studyroomreservation.web.dto.ResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,6 +19,9 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -30,17 +34,17 @@ public class CompanyController {
 
     // Company 여러건 조회
     @GetMapping
-    public ResponseEntity<ResponseDto<Page<CompanyDto>>> getCompanies(Pageable pageable){
-        Page<CompanyDto> companies = companyService.getCompanies(pageable);
-        ResponseDto<Page<CompanyDto>> responseBody = new ResponseDto<>("Success", companies);
+    public ResponseEntity<ResponseDto<Page<CompanyResponseDto>>> getCompanies(Pageable pageable){
+        Page<CompanyResponseDto> companies = companyService.getCompanies(pageable);
+        ResponseDto<Page<CompanyResponseDto>> responseBody = new ResponseDto<>("Success", companies);
         return ResponseEntity.ok().body(responseBody);
     }
 
     // Company 단건 조회
     @GetMapping("/{id}")
-    public ResponseEntity<?> getCompany(@PathVariable("id") Long id){
-        CompanyDto company = companyService.getCompany(id);
-        ResponseDto<CompanyDto> responseBody = new ResponseDto<>("Success", company);
+    public ResponseEntity<ResponseDto<CompanyResponseDto>> getCompany(@PathVariable("id") Long id){
+        CompanyResponseDto company = companyService.getCompany(id);
+        ResponseDto<CompanyResponseDto> responseBody = new ResponseDto<>("Success", company);
         return ResponseEntity.ok().body(responseBody);
     }
 
@@ -52,17 +56,15 @@ public class CompanyController {
      */
     @PutMapping("/{id}")
     public ResponseEntity<?> updateCompany(@PathVariable("id") Long id,
-                                           @RequestBody @Validated CompanyUpdateDto updateDto,
+                                           @RequestPart(value = "files") List<MultipartFile> files,
+                                           @RequestPart(value = "data") @Validated CompanyUpdateDto updateDto,
                                            BindingResult bindingResult,
                                            @AuthenticationPrincipal UserDto userDto){
         // STUDYROOM_ADMIM이면
-        if(UserRole.ROLE_STUDYROOM_ADMIN.name().equals(userDto.getRole().name())){
-            Company company = companyService.findById(id);
-            if(!company.getUser().getId().equals(userDto.getId())){
-                ResponseDto<String> responseBody =
-                        new ResponseDto<>("Access Denied", "No Permission to modify this company");
-                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(responseBody);
-            }
+        if(!checkPermission(userDto, id)){
+            ResponseDto<String> responseBody =
+                    new ResponseDto<>("Access Denied", "No Permission to modify this company");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(responseBody);
         }
 
         if(bindingResult.hasErrors()){
@@ -70,8 +72,9 @@ public class CompanyController {
             ResponseDto<BindingResult> responseBody = new ResponseDto<>("Validation Error", bindingResult);
             return ResponseEntity.badRequest().body(responseBody);
         }
-        CompanyDto companyDto = companyService.updateCompany(id, companyMapper.requestToDto(updateDto));
-        ResponseDto<CompanyDto> responseBody = new ResponseDto<>("Success", companyDto);
+        CompanyUpdateResponseDto responseDto = companyService.updateCompany(id, companyMapper.updateToDto(updateDto), files, updateDto.getDeleteImages());
+
+        ResponseDto<CompanyUpdateResponseDto> responseBody = new ResponseDto<>("Success", responseDto);
         return ResponseEntity.ok().body(responseBody);
     }
 
@@ -86,16 +89,24 @@ public class CompanyController {
     public ResponseEntity<ResponseDto<?>> deleteCompany(@PathVariable("id") Long id,
                                                              @AuthenticationPrincipal UserDto userDto){
         // STUDYROOM_ADMIM이면
-        if(UserRole.ROLE_STUDYROOM_ADMIN.name().equals(userDto.getRole().name())){
-            Company company = companyService.findById(id);
-            if(!company.getUser().getId().equals(userDto.getId())){
-                ResponseDto<String> responseBody =
-                        new ResponseDto<>("Access Denied", "No Permission to modify this company");
-                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(responseBody);
-            }
+        if(!checkPermission(userDto, id)){
+            ResponseDto<String> responseBody =
+                    new ResponseDto<>("Access Denied", "No Permission to modify this company");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(responseBody);
         }
+
         companyService.deleteCompany(id);
         ResponseDto<String> responseBody = new ResponseDto<>("Success", "Delete Company id = " + id);
         return ResponseEntity.ok().body(responseBody);
+    }
+
+    private Boolean checkPermission(UserDto userDto, Long companyId){
+        if (UserRole.ROLE_STUDYROOM_ADMIN.name().equals(userDto.getRole().name())) {
+            Company company = companyService.findById(companyId);
+            if (!company.getUser().getId().equals(userDto.getId())) {
+                return false;
+            }
+        }
+        return true;
     }
 }
